+++ run_locally.sh
@@ -5,6 +5,7 @@ docker network create nodejs_tracing
 docker stop jaeger
 docker stop tracing-nodejs
 
+# 16686
 docker run --rm -d --net nodejs_tracing --name jaeger \
 -e \
   COLLECTOR_ZIPKIN_HTTP_PORT=9411 \
@@ -17,4 +18,11 @@ docker run --rm -d --net nodejs_tracing --name jaeger \
   -p 9411:9411 \
   jaegertracing/all-in-one:latest
 
-docker run --rm -it --net nodejs_tracing --name tracing-nodejs -p 8000:8000 jecnua/tracing-nodejs:dev-latest
+docker run --rm -d --net nodejs_tracing --name tracing-nodejs -p 8000:8000 jecnua/tracing-nodejs:dev-latest
+
+sleep 5
+
+curl localhost:8000
+curl localhost:8000/home
+curl localhost:8000/about
+curl localhost:8000/error